Date: Sun, 27 Jul 2003 22:37:59 +0200 From: Max Horn Subject: SDL whitespace patch
this patch is lowest priority: it only contains indention fixes. In
some places, instead of tabs, whitespaces where used for indention. But
of course that only "looks good" if you use the exact same tab width as
the author of those files. I tried to replace such cases with tabs for
a more uniform indention. I only did that for a few files.

So, no code changes in this, only whitespaces transformed.


git-svn-id: http://svn.libsdl.org/trunk/SDL@1034 c70aab31-4412-0410-b14c-859654838e24