Date: Mon, 17 Nov 2003 23:12:59 +0100 From: Max Horn Subject: SDL building "the unix way" on OS X
I just noticed that building current SDL CVS "the unix way" (using
configure/make) is (and has been for some time, it seems) broken.
That's because Sam updated to a newer libtool version, which removed my
"-framework" patches.

Attached is a patch which once again makes ltmain.sh aware of
"-framework". Note that this is simply a backport of the corresponding
changes in libtool CVS (i.e. the next libtool release will contain
those same changes).


git-svn-id: http://svn.libsdl.org/trunk/SDL@1158 c70aab31-4412-0410-b14c-859654838e24