Date: Thu, 24 Jun 2004 15:13:40 +0300 (EEST) From: Martin_Storsj� Subject: Small patches for SDL/win32
I've tried playing with running SDL compiled for Windows with Wine (and
WineX) under Linux. It didn't originally work too well, but I somehow
managed to make it work. Most of the fixes was patches to Wine(X), but I
also found something in video/windx5/SDL_dx5video.c which, as far as I can
tell, actually is a bug in SDL. The result from the COM method Release,
which returns the new reference count, is compared to DD_OK. Because
QueryInterface is called immediately before, the reference count shouldn't
be DD_OK (== 0).


git-svn-id: http://svn.libsdl.org/trunk/SDL@1357 c70aab31-4412-0410-b14c-859654838e24