Date: Sun, 25 Jul 2004 23:10:03 +0800 From: Chris Taylor Subject: SDL for Macintosh Programmer's Workshop: removed dynamic loading
I'm working on a game called D2X which uses SDL, and I recently ported
it to Mac OS 9. I used MPW to build it. It uses OpenGL. To get it to
work, I had to build SDL so it doesn't dynamically load OpenGL. This is
because I don't think MPW supports dynamic loading. The following patch
makes sure when SDL is built under MPW, dynamic loading of shared
libraries is disabled.


git-svn-id: http://svn.libsdl.org/trunk/SDL@1374 c70aab31-4412-0410-b14c-859654838e24