Date: Mon, 30 Aug 2004 18:20:25 +0200 From: Joost Baas Subject: why call arts artsc?
I, and a few other people at the mplayer-docs-mailinglist were wondering why
you decided to call arts artsc. I understand usually users have nothing to do
with libsdl, just developers, but because you can choose the audio-driver
being used by mplayer, one of which is sdl, and you can also choose the sdl
subdriver, it is necessary to have a well-known or logical name. artsc is not
the logical choice, and it's very hard to look up the right name if you don't
know what you're looking for.


git-svn-id: http://svn.libsdl.org/trunk/SDL@1425 c70aab31-4412-0410-b14c-859654838e24