Date: Sat, 9 Oct 2004 02:46:18 +0300 (EEST) From: Hannu Savolainen Subject: SDL uses obsolete OSS features
I was trying to compile SDL against the latest OSS version. It appeared
that SDL 1.2.7 is using a very old SOUND_PCM_WRITE_RATE ioctl that is
obsolete and has been removed from OSS. The right ioctl to call is
SNDCTL_DSP_SPEED. Attached is a patch (works with all OSS versions).


git-svn-id: http://svn.libsdl.org/trunk/SDL@1431 c70aab31-4412-0410-b14c-859654838e24