Date: Tue, 5 Oct 2004 11:04:02 -0400 From: Bob Ippolito Subject: [SDL] PATCH: Fix for release count bug in SDL_QuartzWM
SDL_QuartzWM incorrectly does [window close]; [window release];

The semantics for -[NSWindow close] are as follows:

Removes the receiver from the screen. If the receiver is set to be
released when closed, a release message is sent to the object after the
current event is completed. For an NSWindow object, the default is to
be released on closing, while for an NSPanel object, the default is not
to be released. You can use the setReleasedWhenClosed: method to change
the default behavior.

This patch simply removes the incorrect release.


git-svn-id: http://svn.libsdl.org/trunk/SDL@1438 c70aab31-4412-0410-b14c-859654838e24