Date: Sun, 2 Jan 2005 18:36:09 +0100 From: Jon Daniel Subject: [SDL] KeyRepeat fix
If you switch keys very fast and hold the last one it
doesn't get repeated because the key release of the first key
occured after the key press of the second key.
Unfortunatly the key release of the first key sets
SDL_KeyRepeat.timestamp = 0; causing the the second key not to repeat.

This should be always reproducable. I've only checked this on x86-64
X11.

To fix this I just added another condition to make sure the sym of the
released key matches SDL_KeyRepeat.key.keysym.sym.


git-svn-id: http://svn.libsdl.org/trunk/SDL@1507 c70aab31-4412-0410-b14c-859654838e24