Date: Wed, 17 Aug 2005 11:23:40 -0400 From: Matt L <prometheus.uw@gmail.com> To: "Ryan C. Gordon" <icculus@clutteredmind.org> Subject: SDL Patch
Hio,

I saw your last call on the mailing list. Here's a patch which I
submitted about two weeks ago which hasn't made it in.

In the current sdl.m4, there's a bug where if your configure.ac, you
have AC_LANG(C++) specified, it won't properly compile and link the SDL
test program when you run the configure script. This is because only the
default CFLAGS is overriden in sdl.m4, and in the patch below, I've
fixed it so it overrides CXXFLAGS as well, allowing it to work with g++.


git-svn-id: http://svn.libsdl.org/trunk/SDL@1633 c70aab31-4412-0410-b14c-859654838e24