Date: Mon, 22 Aug 2005 04:22:46 -0400 From: "Ryan C. Gordon" Subject: [Fwd: SDL patch: make usage of rpath optional]
Here's a small patch against current SDL CVS that makes usage of rpath
optional, by passing --disable-rpath to configure.  This comes in handy
when redistributing SDL -- the rpath setting prevents the lib being
loaded if it's not in the rpath, which makes redistributing in packages
to be installed by non-root users pretty much useless.


git-svn-id: http://svn.libsdl.org/trunk/SDL@1646 c70aab31-4412-0410-b14c-859654838e24