From: Tyler Montbriand <tsm@accesscomm.ca> To: sdl@libsdl.org Date: Fri, 30 Sep 2005 02:24:50 -0600 Subject: [SDL] WinCE timers, continued
Here's a strange timer for Windows CE that doesn't ignore time across
suspends.  It uses GetSystemTime to keep the time continuous, and GetTicks to
get finer-grained readings than 1 second.  It detects the difference between
the GetTicks time and GetSystemTime time on power-on to keep the error within
one second max.

It's not a patch on the current win32 timer code -- took one look at that and
figured it had more than enough #ifdefs already.  It's windows-ce specific.

Another thing I've noticed is that the Windows CE 4.0 and newer API has
functions warn processes about suspends.  This is something SDL REALLY needs
for audio in particular, because turning it off while it's playing causes
anything that uses audio to hardlock the system on power-on.  Unfortunately I
don't have 4.0 to play with.  :(


git-svn-id: http://svn.libsdl.org/trunk/SDL@1707 c70aab31-4412-0410-b14c-859654838e24