Date: Sun, 11 Dec 2005 20:37:04 +0100 From: Olivier Boudeville <olivier.boudeville@online.fr> To: "A list for developers using the SDL library. \(includes SDL-announce\)" <sdl@libsdl.org> Subject: [SDL] NetBSD build patch
Hi everybody,

apparently the SDL-1.2.9 source archive could not compile "as is" on
NetBSD 2.0_STABLE due to a pthread detection issue in the configure script.

I attached a small patch that can be applied to configure.in so that SDL
can be directly (i.e. without the NetBSD package manager) configured and
built successfully on NetBSD (at least on the one I tried !).

Hope this helps,

Olivier.


git-svn-id: http://svn.libsdl.org/trunk/SDL@1732 c70aab31-4412-0410-b14c-859654838e24