Date: Tue, 21 Mar 2006 15:47:44 -0800 From: Eric Wing Subject: Re: Updated tarball
I finally got everything together. Attached is the new framework
package, new Dev-Extra's package, and the updated Xcode project.

The framework pacakge is pretty much unchanged except for the new
binary.

The Xcode project has undergone some major clean up based on the new
changes. I have also removed the exports list  and we'll see what
happens.

The optional developer-extra's package has been overhauled and now
builds a .dmg instead of a .pkg.

The SDLTest project has undergone some clean up as well plus updates
for the new tests.

I also redid the Xcode project templates. The old Project Builder
templates required people to manually convert "to a native target"
which was messy. I also make the new templates copy (embed) the
SDL.framework into the bundle automatically making it potentially
drag-and-drop ready. I also needed to apply a few code fixes to one
of the SDLMain.m's.

By the way, the SDLMain in the main repository is totally out of sync
with the project template ones. I copied the main repository version
in to the SDL App and OpenGL App templates, but somebody should look
into merging the recent SDLMain changes into the CocoaMenu and
CustomCocoa versions of the SDLMain's. Maybe some of this code should
be placed inside the main repository so it is not overlooked as it is
now being inside an Xcode tar-ball.

I also added a bunch of documentation too.


git-svn-id: http://svn.libsdl.org/trunk/SDL@2133 c70aab31-4412-0410-b14c-859654838e24