Fixed bug #177 The attached patch changes configure.in to use "-framework FOO" instead of "-Wl,-framework,FOO".
This avoid some issues with certain versions of libtool that do not handle
-framework properly.
Some versions of libtool will try to reorder the two parts of the option, or
render the 2nd part of the argument to the relative path of a non-existent
library.

Note: It is not enough if SDL uses a version of libtool that does this
correctly, because these -framework options show up in "sdl-config --libs".
Hence, some 3rd party apps which still ship with an old libtool have troubles
compiling under Mac OS X.


git-svn-id: http://svn.libsdl.org/trunk/SDL@2168 c70aab31-4412-0410-b14c-859654838e24