Fixed bug #193
The attached patch is mostly cosmetic: Currently, we always add both -framework
OpenGL and -framework AGL for Mac OS X command line builds.

However, the former is only used for the Cocoa video driver, and the latter
only for the Carbon video driver (as far as I can tell, at least). Hence the
attached patch modifies configure.in so that each only gets added to the list
of frameworks if the corresponding video driver is enabled.


git-svn-id: http://svn.libsdl.org/trunk/SDL@2183 c70aab31-4412-0410-b14c-859654838e24