Date: Wed, 10 May 2006 17:55:51 -0700 From: Eric Wing Subject: Re: Updated tarball
Okay, here's an updated Xcode tarball. I found and purged the
remaining useless binaries so the file is much smaller. This contains
updated copies of the SDLMain.m for the project templates. You still
need to commit the master SDLMain.m (from the previous email) to the
repo though.

I also switched back the PPC SDK to 10.4u. I'm actually not sure what
happens if it is set to 10.2 and the user didn't install it. Maybe
Xcode is smart enough to handle it, but I'm not sure I want to push
it. I'm worried it might handle it in a way that is worse.


git-svn-id: http://svn.libsdl.org/trunk/SDL@2378 c70aab31-4412-0410-b14c-859654838e24