Reverted revision 3416, fixing bug #589 http://bugzilla.libsdl.org/show_bug.cgi?id=589
I was going to add the current window to the OpenGL context info, but that
doesn't fix the case where you set the current context to NULL and then set
the current context to the same window it had before.

This also doesn't take into account changes to the window that might affect
the context, such as viewport changing or fullscreen/windowed mode changing.

Any ideas?


git-svn-id: http://svn.libsdl.org/trunk/SDL@3622 c70aab31-4412-0410-b14c-859654838e24